Desafio: Aplicando o conhecimento em REST para Estado.

+++ src/main/java/com/algaworks/algafoodapi/infrastructure/repository/EstadoRepositoryImpl.java
@@ -3,9 +3,12 @@ package com.algaworks.algafoodapi.infrastructure.repository;
 import com.algaworks.algafoodapi.domain.model.Estado;
 import com.algaworks.algafoodapi.domain.repository.EstadoRepository;
 import java.util.List;
+import java.util.Objects;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import javax.persistence.TypedQuery;
+
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -38,9 +41,12 @@ public class EstadoRepositoryImpl implements EstadoRepository
 
 	@Override
 	@Transactional
-	public void remover(Estado estado)
+	public void remover(Long id)
 	{
-		estado = buscarPorId(estado.getId());
+		Estado estado = buscarPorId(id);
+		if (Objects.isNull(estado)){
+			throw new EmptyResultDataAccessException(1);
+		}
 		entityManager.remove(estado);
 	}
 }
